document_processor/requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
albucore==0.0.23
albumentations==2.0.5
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.2
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
easyocr==1.7.2
exceptiongroup==1.2.2
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
flake8==7.1.1
flake8-html==0.4.3
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.2.0
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
keras==3.9.2
kiwisolver==1.4.8
langchain==0.3.24
langchain-community==0.3.22
langchain-core==0.3.55
langchain-google-genai==2.0.10
langchain-openai==0.3.14
langchain-text-splitters==0.3.8
langsmith==0.3.33
lazy_loader==0.4
libclang==18.1.1
llamaapi==0.1.36
lmdb==1.6.2
lxml==5.4.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
namex==0.0.9
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
numpy==2.1.3
openai==1.76.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt-einsum==3.3.0
optree==0.15.0
orjson==3.10.16
packaging==24.2
paddle-bfloat==0.1.7
paddleocr==2.10.0
paddlepaddle==2.5.0
pdf2image==1.17.0
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyclipper==1.3.0.post6
pycodestyle==2.12.1
pydantic
pydantic-settings==2.9.1
pydantic_core==2.33.1
pyflakes==3.2.0
Pygments==2.18.0
pymongo==4.11.1
pyparsing==3.2.3
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9.1
safetensors==0.5.2
scikit-image==0.25.2
scipy==1.15.2
shapely==2.1.0
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
stringzilla==3.12.5
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0
